+++ module-apps/application-desktop/widgets/PinLockHandler.cpp
@@ -8,7 +8,6 @@
 #include "application-desktop/windows/Names.hpp"
 
 #include <module-utils/common_data/EventStore.hpp>
-
 #include <service-appmgr/service-appmgr/data/SimActionsParams.hpp>
 #include <service-cellular/CellularMessage.hpp>
 
@@ -22,197 +21,252 @@ namespace gui
         constexpr unsigned int blocked_sim_attempts    = 0;
         constexpr unsigned int sim_max_passcode_size   = 8;
         constexpr unsigned int sim_min_passcode_size   = 4;
+
+        constexpr inline auto serviceCellular = "ServiceCellular";
     } // namespace
 
     PinLockHandler::PinLockHandler(app::ApplicationDesktop *app, SettingsRecord &settings)
-        : app(app), appSettings(settings), screenLock(PinLock::SimCard::NoCard,
-                                                      PinLock::LockState::PasscodeRequired,
-                                                      PinLock::LockType::Screen,
-                                                      default_attempts),
-          simLock(PinLock::SimCard::NoCard, PinLock::LockState::Unlocked, PinLock::LockType::SimPin, default_attempts)
+        : app(app), appSettings(settings),
+          screenLock(
+              Store::GSM::SIM::NONE, PinLock::LockState::PasscodeRequired, PinLock::LockType::Screen, default_attempts),
+          simLock(Store::GSM::SIM::NONE, PinLock::LockState::Unlocked, PinLock::LockType::SimPin, default_attempts)
     {
-        reloadScreenLock();
-        screenLock.setAutoActivate(true);
-
         simLock.setPinSizeBounds(sim_min_passcode_size, sim_max_passcode_size);
+        screenLock.setAutoActivate(true);
     }
 
     void PinLockHandler::handleScreenPin(const std::vector<unsigned int> &pin)
     {
-        if (screenLock.getMaxPinSize() == screen_nopin_size) {
+        std::hash<std::vector<unsigned int>> hashEngine;
+        uint32_t hash = hashEngine(pin);
+        screenLock.value--;
+        if (hash == appSettings.lockPassHash) {
             screenLock.lockState = gui::PinLock::LockState::Unlocked;
+            screenLock.value     = default_attempts;
         }
         else if (screenLock.value > 0) {
-            std::hash<std::vector<unsigned int>> hashEngine;
-            uint32_t hash = hashEngine(pin);
-            screenLock.value--;
-            if (hash == appSettings.lockPassHash) {
-                screenLock.lockState = gui::PinLock::LockState::Unlocked;
-                screenLock.value     = default_attempts;
-            }
-            else if (screenLock.value > 0) {
-                screenLock.lockState = gui::PinLock::LockState::PasscodeInvalidRetryRequired;
-            }
-            else {
-                screenLock.lockState = gui::PinLock::LockState::Blocked;
-            }
-        }
-    }
-
-    void PinLockHandler::handleScreenConfirm()
-    {
-        simLock.consumeState();
-    }
-
-    void PinLockHandler::reloadScreenLock()
-    {
-        screenLock.lockState = gui::PinLock::LockState::PasscodeRequired;
-        unsigned int pinSize = appSettings.lockPassHash == 0 ? screen_nopin_size : default_screen_pin_size;
-        screenLock.setPinSizeBounds(pinSize, pinSize);
-        screenLock.value = default_attempts;
-    }
-
-    auto PinLockHandler::handlePasscodeParams(app::manager::actions::ActionParamsPtr &&data) const -> gui::PinLock
-    {
-        auto passcodeData = static_cast<app::manager::actions::PasscodeParams *>(data.get());
-
-        gui::PinLock lock(parseSimCard(passcodeData->getSim()),
-                          gui::PinLock::LockState::PasscodeRequired,
-                          simLock.getLockType(),
-                          passcodeData->getAttempts());
-
-        if (simLock.getLockType() == lock.getLockType() && simLock.getState() == lock.getState() &&
-            simLock.getValue() > lock.getValue()) {
-            lock.lockState = PinLock::LockState::PasscodeInvalidRetryRequired;
-        }
-        lock.setPinSizeBounds(sim_min_passcode_size, sim_max_passcode_size);
-
-        return lock;
-    }
-
-    void PinLockHandler::handlePinRequest(app::manager::actions::ActionParamsPtr &&data)
-    {
-        simLock          = handlePasscodeParams(std::move(data));
-        simLock.lockType = gui::PinLock::LockType::SimPin;
-
-        auto onActivatedCallback = [this](const std::vector<unsigned int> &data) { handlePasscode(data); };
-        switchToPinlockWindow(onActivatedCallback);
-    }
-
-    void PinLockHandler::handlePukRequest(app::manager::actions::ActionParamsPtr &&data)
-    {
-        simLock          = handlePasscodeParams(std::move(data));
-        simLock.lockType = gui::PinLock::LockType::SimPuk;
-
-        auto onActivatedCallback = [this](const std::vector<unsigned int> &data) { handlePasscodeChange(data); };
-        switchToPinlockWindow(onActivatedCallback);
-    }
-
-    void PinLockHandler::switchToPinlockWindow(
-        std::function<void(const std::vector<unsigned int> &)> onLockActivatedCallback)
-    {
-        auto lock = std::make_unique<gui::PinLock>(simLock);
-
-        if (lock->getState() == PinLock::LockState::PasscodeInvalidRetryRequired) {
-            lock->onActivatedCallback = [this, onLockActivatedCallback](const std::vector<unsigned int> &) {
-                simLock.consumeState();
-                switchToPinlockWindow(onLockActivatedCallback);
-            };
+            screenLock.lockState = gui::PinLock::LockState::PasscodeInvalidRetryRequired;
         }
         else {
-            lock->onActivatedCallback = onLockActivatedCallback;
+            screenLock.lockState = gui::PinLock::LockState::Blocked;
         }
+        unlock();
+    }
 
-        app->switchWindow(app::window::name::desktop_pin_lock,
-                          gui::ShowMode::GUI_SHOW_INIT,
-                          std::make_unique<gui::LockPhoneData>(std::move(lock)));
+    void PinLockHandler::handlePasscodeParams(PinLock::LockType type,
+                                              PinLock::LockState state,
+                                              app::manager::actions::ActionParamsPtr &&data)
+    {
+        auto passcodeData = static_cast<app::manager::actions::PasscodeParams *>(data.get());
+        if (simLock.isSim(passcodeData->getSim()) && simLock.isType(type) && simLock.isState(state) &&
+            simLock.getValue() > passcodeData->getAttempts()) {
+            simLock.lockState = PinLock::LockState::PasscodeInvalidRetryRequired;
+        }
+        else {
+            simLock.lockState = state;
+            simLock.sim       = passcodeData->getSim();
+            simLock.lockType  = type;
+        }
+        simLock.value = passcodeData->getAttempts();
+    }
+
+    void PinLockHandler::handlePasscodeRequest(PinLock::LockType type, app::manager::actions::ActionParamsPtr &&data)
+    {
+        LOG_DEBUG("Handling on of PasscodeRequest actions");
+        handlePasscodeParams(type, PinLock::LockState::PasscodeRequired, std::move(data));
+        if (!getStrongestLock().isType(PinLock::LockType::Screen)) {
+            unlock();
+        }
     }
 
     void PinLockHandler::handlePinChangeRequest(app::manager::actions::ActionParamsPtr &&data)
     {
-        simLock          = handlePasscodeParams(std::move(data));
-        simLock.lockType = gui::PinLock::LockType::SimPin;
-
-        auto onActivatedCallback = [this](const std::vector<unsigned int> &data) { handlePasscodeChange(data); };
-        switchToPinlockWindow(onActivatedCallback);
+        LOG_DEBUG("Handling RequestPinChange action");
+        handlePasscodeParams(PinLock::LockType::SimPin, PinLock::LockState::PasscodeRequired, std::move(data));
+        promptSimLockWindow      = true;
+        auto onActivatedCallback = [this](PinLock::LockType type, const std::vector<unsigned int> &data) {
+            handlePasscodeChange(data);
+        };
+        switchToPinLockWindow(onActivatedCallback);
     }
 
     void PinLockHandler::handleSimBlocked(app::manager::actions::ActionParamsPtr &&data)
     {
-        auto params = static_cast<app::manager::actions::SimStateParams *>(data.get());
-        simLock     = PinLock(parseSimCard(params->getSim()),
-                          gui::PinLock::LockState::Blocked,
-                          gui::PinLock::LockType::SimPin,
-                          blocked_sim_attempts);
+        LOG_DEBUG("Handling BlockSim action");
+        auto params         = static_cast<app::manager::actions::SimStateParams *>(data.get());
+        promptSimLockWindow = true;
+        simLock.sim         = params->getSim();
+        simLock.lockState   = gui::PinLock::LockState::Blocked;
+        simLock.lockType    = gui::PinLock::LockType::SimPin;
+        simLock.value       = blocked_sim_attempts;
+        unlock();
+    }
 
-        auto lock = std::make_unique<gui::PinLock>(simLock);
-        app->switchWindow(app::window::name::desktop_pin_lock,
-                          gui::ShowMode::GUI_SHOW_INIT,
-                          std::make_unique<gui::LockPhoneData>(std::move(lock)));
+    void PinLockHandler::handleUnlockSim(app::manager::actions::ActionParamsPtr &&data)
+    {
+        LOG_DEBUG("Handling UnlockSim action");
+        simLock.lockState   = PinLock::LockState::Unlocked;
+        promptSimLockWindow = false;
+        unlock();
     }
 
     void PinLockHandler::handleCMEError(app::manager::actions::ActionParamsPtr &&data) const
     {
+        LOG_DEBUG("Handling DisplayCMEError action");
         auto params = static_cast<app::manager::actions::UnhandledCMEParams *>(data.get());
-
-        auto lock = std::make_unique<gui::PinLock>(parseSimCard(params->getSim()),
-                                                   PinLock::LockState::ErrorOccurred,
-                                                   PinLock::LockType::SimPin,
-                                                   params->getCMECode());
-
+        auto lock   = std::make_unique<gui::PinLock>(
+            params->getSim(), PinLock::LockState::ErrorOccurred, PinLock::LockType::SimPin, params->getCMECode());
+        lock->onActivatedCallback = [this](PinLock::LockType type, const std::vector<unsigned int> &data) {
+            app->switchWindow(app::window::name::desktop_main_window);
+        };
         app->switchWindow(app::window::name::desktop_pin_lock,
                           gui::ShowMode::GUI_SHOW_INIT,
                           std::make_unique<gui::LockPhoneData>(std::move(lock)));
     }
 
-    void PinLockHandler::handlePasscode(const std::vector<unsigned int> passcode) const
+    void PinLockHandler::switchToPinLockWindow(
+        std::function<void(PinLock::LockType, const std::vector<unsigned int> &)> onLockActivatedCallback)
     {
-        if (simLock.getLockType() == PinLock::LockType::SimPin) {
-            sys::Bus::SendUnicast(
-                std::make_shared<CellularSimPinDataMessage>(Store::GSM::SIM::SIM1, passcode), "ServiceCellular", app);
+        auto lock = std::make_unique<gui::PinLock>(getStrongestLock());
+        if (lock->isState(PinLock::LockState::PasscodeInvalidRetryRequired)) {
+            lock->onActivatedCallback = [this, onLockActivatedCallback](PinLock::LockType,
+                                                                        const std::vector<unsigned int> &) {
+                getStrongestLock().consumeState();
+                switchToPinLockWindow(onLockActivatedCallback);
+            };
         }
-        else if (simLock.getLockType() == PinLock::LockType::SimPuk) {
+        else if (lock->isState(gui::PinLock::LockState::Blocked)) {
+            lock->onActivatedCallback = [this](PinLock::LockType type, const std::vector<unsigned int> &data) {
+                setSimLockHandled();
+                app->switchWindow(app::window::name::desktop_main_window);
+            };
+        }
+        else if (lock->isState(gui::PinLock::LockState::Unlocked)) {
+            setSimLockHandled();
+            app->switchWindow(app::window::name::desktop_main_window);
+            return;
+        }
+        else {
+            lock->onActivatedCallback = onLockActivatedCallback;
+        }
+        app->switchWindow(app::window::name::desktop_pin_lock,
+                          gui::ShowMode::GUI_SHOW_INIT,
+                          std::make_unique<gui::LockPhoneData>(std::move(lock)));
+    }
+
+    void PinLockHandler::switchToPinLockWindow(
+        PinLock::LockState state,
+        std::function<void(PinLock::LockType, const std::vector<unsigned int> &)> onLockActivatedCallback)
+    {
+        auto lock                 = std::make_unique<gui::PinLock>(getStrongestLock());
+        lock->lockState           = state;
+        lock->onActivatedCallback = onLockActivatedCallback;
+        app->switchWindow(app::window::name::desktop_pin_lock,
+                          gui::ShowMode::GUI_SHOW_INIT,
+                          std::make_unique<gui::LockPhoneData>(std::move(lock)));
+    }
+
+    void PinLockHandler::handlePasscode(PinLock::LockType type, const std::vector<unsigned int> passcode)
+    {
+        if (type == PinLock::LockType::SimPin) {
+            setSimLockHandled();
+            sys::Bus::SendUnicast(
+                std::make_shared<CellularSimPinDataMessage>(simLock.sim, passcode), serviceCellular, app);
+        }
+        else if (type == PinLock::LockType::SimPuk) {
             handlePasscodeChange(passcode);
         }
+        else if (type == PinLock::LockType::Screen) {
+            handleScreenPin(passcode);
+        }
     }
-    void PinLockHandler::handlePasscodeChange(const std::vector<unsigned int> passcode) const
-    {
-        auto lock       = std::make_unique<gui::PinLock>(simLock);
-        lock->lockState = PinLock::LockState::NewPasscodeRequired;
 
-        lock->onActivatedCallback = [this, passcode](const std::vector<unsigned int> &pin) {
-            handleNewPasscodeConfirmed(passcode, pin);
+    void PinLockHandler::handlePasscodeChange(const std::vector<unsigned int> passcode)
+    {
+        auto onActivatedCallback = [this, passcode](PinLock::LockType, const std::vector<unsigned int> &pin) {
+            handleNewPasscodeUnconfirmed(passcode, pin);
         };
-
-        app->switchWindow(app::window::name::desktop_pin_lock,
-                          gui::ShowMode::GUI_SHOW_INIT,
-                          std::make_unique<gui::LockPhoneData>(std::move(lock)));
+        switchToPinLockWindow(PinLock::LockState::NewPasscodeRequired, onActivatedCallback);
     }
 
-    void PinLockHandler::handleNewPasscodeConfirmed(const std::vector<unsigned int> &passcode,
-                                                    const std::vector<unsigned int> &pin) const
+    void PinLockHandler::handleNewPasscodeUnconfirmed(const std::vector<unsigned int> &passcode,
+                                                      const std::vector<unsigned int> &pin)
     {
-        if (simLock.getLockType() == PinLock::LockType::SimPin) {
-            sys::Bus::SendUnicast(std::make_shared<CellularSimNewPinDataMessage>(Store::GSM::SIM::SIM1, passcode, pin),
-                                  "ServiceCellular",
-                                  app);
+        auto onActivatedCallback = [this, passcode, pin](PinLock::LockType type,
+                                                         const std::vector<unsigned int> &pinConfirmed) {
+            if (pin == pinConfirmed) {
+                handleNewPasscodeConfirmed(type, passcode, pin);
+            }
+            else {
+                handleNewPasscodeInvalid(passcode);
+            }
+        };
+        switchToPinLockWindow(PinLock::LockState::NewPasscodeConfirmRequired, onActivatedCallback);
+    }
+
+    void PinLockHandler::handleNewPasscodeConfirmed(PinLock::LockType type,
+                                                    const std::vector<unsigned int> &passcode,
+                                                    const std::vector<unsigned int> &pin)
+    {
+        setSimLockHandled();
+        if (type == PinLock::LockType::SimPin) {
+            sys::Bus::SendUnicast(
+                std::make_shared<CellularSimNewPinDataMessage>(simLock.sim, passcode, pin), serviceCellular, app);
         }
-        else if (simLock.getLockType() == PinLock::LockType::SimPuk) {
-            sys::Bus::SendUnicast(std::make_shared<CellularSimPukDataMessage>(Store::GSM::SIM::SIM1, passcode, pin),
-                                  "ServiceCellular",
-                                  app);
+        else if (type == PinLock::LockType::SimPuk) {
+            sys::Bus::SendUnicast(
+                std::make_shared<CellularSimPukDataMessage>(simLock.sim, passcode, pin), serviceCellular, app);
         }
     }
 
-    auto PinLockHandler::parseSimCard(Store::GSM::SIM sim) const noexcept -> PinLock::SimCard
+    void PinLockHandler::handleNewPasscodeInvalid(const std::vector<unsigned int> &passcode)
     {
-        if (sim == Store::GSM::SIM::SIM1) {
-            return gui::PinLock::SimCard::SIM1;
+        auto onActivatedCallback = [this, passcode](PinLock::LockType type, const std::vector<unsigned int> &pin) {
+            handlePasscodeChange(passcode);
+        };
+        switchToPinLockWindow(PinLock::LockState::NewPasscodeInvalid, onActivatedCallback);
+    }
+    void PinLockHandler::unlock()
+    {
+        auto onActivatedCallback = [this](PinLock::LockType type, const std::vector<unsigned int> &data) {
+            handlePasscode(type, data);
+        };
+        switchToPinLockWindow(onActivatedCallback);
+    }
+
+    auto PinLockHandler::getStrongestLock() noexcept -> gui::PinLock &
+    {
+        if (!screenLock.isState(PinLock::LockState::Unlocked)) {
+            return screenLock;
         }
-        else if (sim == Store::GSM::SIM::SIM2) {
-            return gui::PinLock::SimCard::SIM2;
+        else if (promptSimLockWindow && !simLock.isState(PinLock::LockState::Unlocked)) {
+            return simLock;
         }
-        return gui::PinLock::SimCard::NoCard;
+        return screenLock;
+    }
+
+    void PinLockHandler::lockScreen()
+    {
+        screenLock.lockState = PinLock::LockState::PasscodeRequired;
+    }
+
+    void PinLockHandler::unlockScreen()
+    {
+        if (getStrongestLock().isType(PinLock::LockType::Screen)) {
+            unsigned int pinSize = appSettings.lockPassHash == 0 ? screen_nopin_size : default_screen_pin_size;
+            screenLock.setPinSizeBounds(pinSize, pinSize);
+            if (screenLock.getMaxPinSize() == screen_nopin_size) {
+                screenLock.lockState = gui::PinLock::LockState::Unlocked;
+            }
+        }
+        unlock();
+    }
+
+    void PinLockHandler::setSimLockHandled() noexcept
+    {
+        if (!getStrongestLock().isType(PinLock::LockType::Screen)) {
+            promptSimLockWindow = false;
+        }
     }
 
 } // namespace gui