+++ test/harness/__init__.py
@@ -3,6 +3,6 @@
 
 import logging
 import harness.interface
-# logging.basicConfig(format='%(asctime)s %(message)s', level=logging.DEBUG)
+logging.basicConfig(format='%(asctime)s %(message)s', level=logging.DEBUG)
 log = logging.getLogger(__name__)